DISABLE CHECKSUM FOR WD CLONES, FIX MULTIBOARD ROUTING, ADD 16 BIT XFER
Mike Durkin:
	Danpex EW-2016 is an inexpensive (US$85.) allegedly WD8013-compatible
ethernet card.

I have one of these cards also and after fiddling a bit with the source
found that the checksum in the probe fails.  I commented it out and am
using the card successfully, NFS mounting to and from my Sun3s with no
problems.

This patch reports the failure, but doesn't fail the probe because of it.

[ Editor's note: ]
	Add "option WDCOMPAT" to enable this.  Do NOT do this if your card
already passes the checksum test,  or you don't have a WD80x3
board currently in your system, since disabling the checksum
will result in the kernel finding a board which isn't there.

Michael Galassi:
	Allow routing between multiple Western Digital ethernet boards
which are installed in the same host.

Barry Lustig:
	Enable 16 bit transfers for the WD8013 "Combo board".  This patch
is kernel configuration optional -- that is, it will not take full effect
(other than generic fixes) unless you modify your kernel configuration
file and rerun "config".

[ Editor's note: ]
	Add "option WD8013" to enable this.  Do NOT do this if you will
have two or more wd boards, one of which is not a "Combo".

AUTHOR:	Mike Durkin (mdurkin@tsoft.sf-bay.org)
AUTHOR:	Michael Galassi (nerd@percival.rain.com)
AUTHOR:	Barry Lustig (barry@ictv.com)
386BSD-Patchkit: patch00100