From: jtkohl@decvax.dec.com (John Kohl Base File System)
This bug seems to derive from the Net/2 i386 BSD code; it's in BSD/386,
and probably also in 386BSD.

Description:
        syscall(2) doesn't work right.  It leaves $sp off by a word,
which can lead to disaster when it's called repeatedly (as it is in
XView3.0).

Repeat-By:
        Write a program that has a local variable on the stack;
initialize it, then call syscall() repeatedly.  Note that the value
eventually gets trashed when the stack works its way up over the
variable.