In src/Makefile.inc1:     Remove /sys/boot from legacy-build.     Add btxld to build-tools. In src/sys/Makefile:     Add /sys/boot for i386 ELF.
I'm still not sure why the new boot code was being built along with the
legacy stuff, which meant a completely wrong default environment for it.

This may well still be the wrong way to go about this, but it can't work
all that much worse than it has been.