Significant modifications to the SCSI controller support sections in both files.  They are now both basically the same.  I also modified the driver list in HARDWARE.TXT:  add the adw and isp drivers, and indicate that the uha driver isn't yet supported under CAM.
This includes specific mention of all supported NCR and BusLogic models,
additional qualification of the supported DPT and QLogic models, and some
additions to the list of supported onboard Adaptec chips.

The hope is, of course, that this will lead to fewer questions like "is the
froboz SCSI controller supported?"

I think the formatting of the new entries looks okay, but someone with a
better eye for things like that might want to look at this.