Add the GEOM regression test framework.
This is a set of userland shims in which GEOM can be run through simple
tests.

The simulation of kernel synchronization primitives is very primitive
and consequently some times tests will fail because of races.

Data/ contains a number of files in XML format which describe the
key sectors for a number of disk images

This is a very handy tool for people developing GEOM methods.  The
"simdisk" method can be told to read from a "real disk" and afterwards
dump the accessed sectors in XML format for further use.

I hope future method writes will see the benefit of this test
collection and add to it when they write methods for GEOM.

You will need ports/textproc/expat for the XML parser.

Sponsored by:   DARPA, NAI Labs.