MFC:  src/sys/fs/devfs/devfs_vnops.c		1.128  src/sys/kern/vfs_subr.c		1.652
This is a workaround for a complicated issue involving VFS cookies and devfs.
The PR and patch have the details. The ultimate fix requires architectural
changes and clarifications to the VFS API, but this will prevent the system
from panicking when someone does "ls /dev" while running in a shell under the
linuxulator.

PR:		88249
Submitted by:   "Devon H. O'Dell" <dodell@ixsystems.com>